+++ src/graphics/data_structures/matrix4.test.cpp
@@ -10,7 +10,7 @@ TEST(GraphicsMatrix4, DefaultConstructor)
 
     for (int row = 0; row < 4; ++row)
         for (int col = 0; col < 4; ++col) {
-            const float value = matrix[row, col];
+            const double value = matrix[row, col];
             ASSERT_FLOAT_EQ(value, 0.0);
         }
 }
@@ -18,7 +18,7 @@ TEST(GraphicsMatrix4, DefaultConstructor)
 // Tests the float list constructor
 TEST(GraphicsMatrix4, FloatListConstructor)
 {
-    std::array<const float, 16> matrix_values{
+    std::array<const double, 16> matrix_values{
             1.0, 2.0, 3.0, 4.0,
             5.0, 6.0, 7.0, 8.0,
             9.0, 10.0, 11.0, 12.0,
@@ -33,7 +33,7 @@ TEST(GraphicsMatrix4, FloatListConstructor)
 
     for (int row = 0; row < 4; ++row)
         for (int col = 0; col < 4; ++col) {
-            const float value = matrix[row, col];
+            const double value = matrix[row, col];
             ASSERT_FLOAT_EQ(value, matrix_values[row * 4 + col]);
         }
 }
@@ -41,7 +41,7 @@ TEST(GraphicsMatrix4, FloatListConstructor)
 // Tests the span-based constructor
 TEST(GraphicsMatrix4, SpanConstructor)
 {
-    std::array<const float, 16> matrix_values{
+    std::array<const double, 16> matrix_values{
             1.0, 2.0, 3.0, 4.0,
             5.0, 6.0, 7.0, 8.0,
             9.0, 10.0, 11.0, 12.0,
@@ -51,7 +51,7 @@ TEST(GraphicsMatrix4, SpanConstructor)
 
     for (int row = 0; row < 4; ++row)
         for (int col = 0; col < 4; ++col) {
-            const float value = matrix[row, col];
+            const double value = matrix[row, col];
             ASSERT_FLOAT_EQ(value, matrix_values[row * 4 + col]);
         }
 }
@@ -59,7 +59,7 @@ TEST(GraphicsMatrix4, SpanConstructor)
 // Tests the copy constructor
 TEST(GraphicsMatrix4, CopyConstructor)
 {
-    std::array<const float, 16> matrix_values{
+    std::array<const double, 16> matrix_values{
             1.0, 2.0, 3.0, 4.0,
             5.0, 6.0, 7.0, 8.0,
             9.0, 10.0, 11.0, 12.0,
@@ -70,7 +70,7 @@ TEST(GraphicsMatrix4, CopyConstructor)
 
     for (int row = 0; row < 4; ++row)
         for (int col = 0; col < 4; ++col) {
-            const float value = matrix_b[row, col];
+            const double value = matrix_b[row, col];
             ASSERT_FLOAT_EQ(value, matrix_values[row * 4 + col]);
         }
 }
@@ -78,7 +78,7 @@ TEST(GraphicsMatrix4, CopyConstructor)
 // Tests the assignment operator
 TEST(GraphicsMatrix4, AssignmentOperator)
 {
-    std::array<const float, 16> matrix_values{
+    std::array<const double, 16> matrix_values{
             1.0, 2.0, 3.0, 4.0,
             5.0, 6.0, 7.0, 8.0,
             9.0, 10.0, 11.0, 12.0,
@@ -91,7 +91,7 @@ TEST(GraphicsMatrix4, AssignmentOperator)
 
     for (int row = 0; row < 4; ++row)
         for (int col = 0; col < 4; ++col) {
-            const float value = matrix_b[row, col];
+            const double value = matrix_b[row, col];
             ASSERT_FLOAT_EQ(value, matrix_values[row * 4 + col]);
         }
 }
